Mute track (#49)
## Description

- The `replaceTrack` method can handle a nullable track parameter. A
previously renegotiated track can be stopped and reused in the future.
- The `addTrack` method doesn't require the `stream` parameter. The
FishajmClient creates a new internal `Stream` object for every
`addTrack` invocation, allowing updates to this object. In the previous
implementation, a new track was never added to a stream (likely because
that stream was an external dependency).
- The `localTrackMuted` event is emitted when a track is replaced with
`null`.
- The `localTrackUnmuted` event is emitted when a track is replaced with
a non-null track object.

### E2E tests
To test if this works, I've added tests that check if another peer
receives frames and if the background color changes.

## Motivation and Context

This change allows disabling a media device when it's not in use without
additional renegotiation. Previously, it was necessary to add and remove
the track every time the camera or microphone was started or stopped.

## Types of changes

- [ ] Bug fix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [x] Breaking change (a fix or feature that would cause existing
functionality to not work as expected)
  - The `replaceTrack` method can handle a nullable track parameter.
  - The `stream` parameter was removed from the `addTrack` method.

---------

Co-authored-by: Przemysław Rożnawski <roznawski.przemyslaw@gmail.com>