Mute track (#79)
## Description

This PR introduces changes from:
https://github.com/fishjam-dev/ts-client-sdk/pull/49 and also:

### useCamera, useMicrophone, useScreenshare

Hooks for managing a single track of a given type in the global state
(similarly to the React Native API).

#### onDeviceStop

`onDeviceStop?: "remove" | "mute";
This PR introduces a new `onDeviceStop` config parameter that allows the
user to decide what to do if a track stops. Without it, the track has
been removed from `RTCPeerConnection`. Now, with the option `mute`, that
track can be reused later without renegotiation. It is a companion event
to `onDeviceChange`.

Added `onDeviceStop`, `onDeviceChange`, and the ability to stop a device
to `use-camera-and-microphone-example`.

#### onDeviceChange

`onDeviceChange?: "replace" | "stop";`
Determines whether a track should be replaced when the user requests a
device change. Previously, `broadcastOnDeviceChange`.

#### Other changes

- Names of types like `UseMicrophoneResult` changed to something
simpler, like `MicrophoneAPI`.

## Motivation and Context

Fishjam now offers the ability to reuse tracks.

## Types of changes

- [ ] Bug fix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [x] Breaking change (a fix or feature that would cause existing
functionality to not work as expected)
  - The `replaceTrack` method can handle a nullable track parameter.
  - The `stream` parameter was removed from the `addTrack` method.