Fix simple app example (#124)
## Description

Local storage caused the same token to be set for all browser tabs. This
resulted client B connecting, kicking out the peer A, then peer A tried
to reconnect, and so on.

## Motivation and Context

Why is this change required? What problem does it solve? If it fixes an
open
issue, please link to the issue here.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)