FCE-626 middleware starts two times (#128)
## Description

Logic responsible for managing middleware was moved from
`useTrackManager` to the new class `MiddlewareManager` which is
internally used in `DeviceManager`. Device manager is reposible

## Refactor
- The `storageConfig` from `DeviceManager` has been extracted to
`localStorage`.
- Some functions have been moved to different files to avoid unnecessary
exports.
- `useProcessedPreviewStream.ts`, `useVideoDeviceManager.ts` and
`useVideoDeviceManager.ts` have been removed.

## Motivation and Context

The middleware was not handled correctly. It was executed twice for each
call

## Types of changes

Bug fix (non-breaking change which fixes an issue)