[FCE-1082] - Fix generated output of TrackContextImpl (#259)
## Description

- Refactored TrackContextImpl to avoid runtime errors caused by
useDefineForClassFields.

## Motivation and Context

Our tsconfig either enables `useDefineForClassFields` or sets `target:
es2022` (enabling it by default). This can cause issues when
subclassing, as this may be used before `super()` is called. More
details
[here](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#the-usedefineforclassfields-flag-and-the-declare-property-modifier).
This caused runtime errors with `WebRTCEndpoint`, especially in
Playwright tests for `membrane-rtc-engine`.

We have two options:

1. Disable `useDefineForClassFields`, which is risky and could impact
the codebase.
2. Refactor `TrackContextImpl` using declare to modify JS output and
move initialization to the constructor.

I chose option 2 as it is less invasive and resolves the issue safely.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)