Fce 1279/patch race conditions and poor cleanup (#292)
## Description

This PR mitigates the issue of starting multiple streams and being
unable to stop them later.
The camera is now working correctly in `React.StrictMode`.

Please aware that this is rather **a hack** and a rewrite of the Device
Manager is already planned.

## Motivation and Context

Due to StrictMode nature, the DeviceManager's `start` function was
called twice, leading to 2 streams being retrieved from the browser,
while only one was manageable.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)