Filter tracks based on room type (#301)
## Description

`FishjamClient.addTrack` now checks if the track kind is allowed for the
given room type and throws an error if it's not allowed.
`react-client` handles this error, i.e. the track is not added and a
warning is logged to the console

## Motivation and Context

This aims to give users better feedback when they attempt to add a video
track to an audio-only room, as until now the behavior is that the peer
joins the room and after a few seconds the entire `PeerConnection`
crashes abruptly without any reasonable logs.

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)