Remove `useFishjamClient_DO_NOT_USE` (#305)
## Description

Removes the deprecated hook.
Instead, to satisfy existing e2e tests requirements, I've added a
`useStatistics` hook that exposes a `getStatistics` method, giving
access to low level rtc stats.
If you think this is not needed, we might as well disable the tests that
depend on it, but I'd rather keep them alive.

## Motivation and Context

It's no longer needed.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)