FCE-1453 / Add current device property (#332)
## Description

Introduces `selectedCamera` and `selectedMicrophone` properties.
Deprecates `activeCamera` and `activeMicrophone` properties.

## Motivation and Context

The "active device" property did not provide information about which
device was selected when a device was off.
Right now, "selected device" shows the device that is being used or will
be used after it is turned on.
"Is device on" indicates if the device is turned on.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)